backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
langchain==0.2.16
langchain-text-splitters==0.2.4
langchain-groq==0.1.9
langchain-community==0.2.16
langchain-chroma==0.1.4
chromadb==0.5.23
pydantic==2.10.0
python-dotenv==1.0.0